// api/seller/client/models/UpdateCouponStatusDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { _36_Enums_CouponStatus } from './_36_Enums_CouponStatus';
export type UpdateCouponStatusDto = {
  status: _36_Enums_CouponStatus;
  listCouponId: Array<string>;
};